+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Minor Changes
 - Added the `CoreNamespace.getTokensForOwner()` method to get all the token balances and token metadata for a given address.
+- Added the `wyvern` and `cryptopunks` marketplaces in the `NftSaleMarketplace` enum to the `NftNamespace.getNftSales()` method.
 
 ## 2.4.1