+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Major Changes
 
 ### Minor Changes
+- Added a missing `gas` field in the `DebugTransaction` interface to specify the gas provided for a transaction execution.
 
 ## 2.4.2