+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Minor Changes
 - Added a missing `gas` field in the `DebugTransaction` interface to specify the gas provided for a transaction execution.
+- Added the `BigNumber` ethers export. You can access this by importing `BigNumber` along with the other exports in the package.
 
 ## 2.4.2